% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.ERA20C.R
\name{ERA20C.annual.mean.hourly}
\alias{ERA20C.annual.mean.hourly}
\title{Make annual mean fields for each hour.}
\usage{
ERA20C.annual.mean.hourly(variable, year)
}
\arguments{
\item{variable}{ERA20C variable name, only 2d variables will work.}

\item{year}{Year to make averages for.}
}
\value{
A list (indexed 1:24) of fields, each of which is the annual average for that hour.
}
\description{
Poor man's normals - just the mean for each hour (1:24) over a year
}
\details{
Note that the indices are 1:24 not 0:23 as 0 can't be used as a list index.
}
