% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.ERA20C.R
\name{ERA20C.get.slice.at.hour}
\alias{ERA20C.get.slice.at.hour}
\title{Get slice at hour.}
\usage{
ERA20C.get.slice.at.hour(variable, year, month, day, hour, height = NULL,
  fc.init = NULL, type = "mean")
}
\arguments{
\item{variable}{'prmsl', 'prate', 'air.2m', 'uwnd.10m' or 'vwnd.10m' - or any ERA20C variable.
can also be 'sst' - will return 'air.sfc' over sea grids.}

\item{height}{Height in hPa - leave NULL for monolevel}

\item{fc.init}{If NULL (default) - use the most recent forecast start. If 'last', use the previous day's
forecast start (only possible for 9am data). If 'blend', smooth the transition between forecast fields
initialised on one day and those initialised on the following day by using the 27-hour forcast
on the first day to interpolate.}

\item{type}{- 'mean' (default), 'spread', 'normal', or 'standard.deviation'. 
Note that standard deviations are not available over opendap.}
}
\value{
A GSDF field with lat and long as extended dimensions
}
\description{
Get a 2D horizontal slice of a selected variable (as a GSDF field) for a given hour.
}
\details{
Interpolates to the selected hour when the data available are less than hourly.
Interpolates to the selected height when the selected height is not that of a ERA20C level.
}
