% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ERA20C.hourly.get.file.name}
\alias{ERA20C.hourly.get.file.name}
\title{ERA20C get file name (hourly)}
\usage{
ERA20C.hourly.get.file.name(variable, year, month, day, hour, height = NULL,
  type = "mean")
}
\arguments{
\item{variable}{'prmsl', 'prate', 'air.2m', 'uwnd.10m' or 'vwnd.10m' - or any ERA20C variable}

\item{type}{- 'mean', 'spread', 'normal', or 'standard.deviation'. - must be 'mean'
Note that standard deviations are not available over opendap.}
}
\value{
File name or URL for netCDF file containing the requested data
}
\description{
Get the file name for selected variable and date (hourly data)
}
\details{
Called internally by \code{ERA20C.get.slice.at.hour} but also useful
 called directly - you can then access the data with another tool.
}

