% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.ERA20C.R
\name{ERA20C.get.data.dir}
\alias{ERA20C.get.data.dir}
\title{ERA20C get data directory}
\usage{
ERA20C.get.data.dir()
}
\value{
Base directory name (or NULL is no local files)
}
\description{
Find local data directory - different for different systems
}
\details{
The various different systems on which I run this code all have
 different places to keep large data files. This function returns
 the right base directory for the system (if there is one).
}
