% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSDF.ERA20C.R
\name{ERA20C.show.variables}
\alias{ERA20C.show.variables}
\title{ERA20C show variables}
\usage{
ERA20C.show.variables()
}
\description{
List the variables available from 20CR
}
\details{
2 Different grids: 'Monolevel' is
Lat x Lon x Time, 'Pressure level'
is  Lat x Lon x Height x Time. No parameters or return value
prints out a list of available variables.
}
